Added task 2.2
