+++ package.json
@@ -29,12 +29,15 @@
     "@babel/preset-env": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
     "@types/express": "^4.17.7",
+    "@types/hapi__joi": "^17.1.4",
     "@types/uuid": "^8.3.0",
     "nodemon": "^2.0.4"
   },
   "dependencies": {
+    "@hapi/joi": "^17.1.1",
     "csvtojson": "^2.0.10",
     "express": "^4.17.1",
+    "express-joi-validation": "^4.0.4-beta.0",
     "uuid": "^8.3.0"
   }
 }
